refactor: move renderType function
