--- packages/serverless-typespec-generator/src/typespec/ir/emit.test.ts
@@ -1,10 +1,0 @@
-import { formatTypeSpec } from "@typespec/compiler"
-
-async function normalizeTypeSpec(code: string) {
-  const formattedCode = await formatTypeSpec(code, {
-    indent: "  ",
-    lineWidth: 80,
-    trailingNewline: false,
-  })
-  return formattedCode.trimEnd()
-}
